backend/connectors/app/requirements.txt
fastapi
uvicorn
atlassian-python-api
openpyxl
bidict==0.11.0
sqlalchemy==1.3.20
pandas==1.1.4
psycopg2-binary==2.8.6